+++ worlds/kirbyam/options.py
@@ -91,6 +91,12 @@ class EnableGameplayStateDebugLogging(Toggle):
     default = 0
 
 
+class RoomSanity(Toggle):
+    """Adds room-visit checks (Room X-YY). Disabled by default because it adds 257 locations."""
+    display_name = "Room Sanity"
+    default = 0
+
+
 class KirbyAmDeathLink(DeathLink):
     __doc__ = DeathLink.__doc__
 
@@ -161,6 +167,8 @@ class KirbyAmOptions(PerGameCommonOptions):
     randomize_boss_spawned_ability_grants: RandomizeBossSpawnedAbilityGrants
 
     randomize_miniboss_ability_grants: RandomizeMiniBossAbilityGrants
+
+    room_sanity: RoomSanity
 
     enable_gameplay_state_debug_logging: EnableGameplayStateDebugLogging